requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
scipy>=1.11.0
numpy>=1.24.0
